FEAT: Add flag to ignore empty IPs
If you, for example, have a server behind a router that regularly
restarts weetee would not reach the "what's my ip" endpoint and record
the new IP as "" (no IP). This would then be sent as a notification of
an IP change. Then when the server can reach the endpoint again it would
send another email reporting that it changed from IP "" to the initial
IP. Other scenarios may result in a similar behavior, for example if the
"what's my ip" endpoint goes down.

This can result in a decent amount of "spam" or at least unactionable
and not very useful notifications.

Since the default behavior have been to send these messages we allow
users to opt-in to the new behavior. potentially one could use the
default behavior as a check that something restarted but you shouldn't
need to get those emails.